result-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
confluent-kafka==2.5.3
psycopg2-binary==2.9.9
pydantic==2.8.2
